Change quoins for Rainbow Run to be currants to match video evidence
Quoins for Rainbow Run are defaulted to 'mood' which does not match
the video video evidence
(https://www.dropbox.com/sh/3t3iqmj7au52xmd/AAA8rFXUHrPVmEx2NiJaWcK-a
/Video/Rainbow%20Run?dl=0&preview=Rainbow+Run.mp4).
As these quoins respawn in under 25s, and are clearly currant quoins,
the simplest fix is to change the type setting.

Note that this fix cannot be tested because Rainbow Run does not yet
exist in fixtures.