Fix visiting stones
* Instead of making webapp callbacks, a new group is used to store all
player TSIDs that have opted in. This is basic right now and doesn't
check if the target is the current player or send multiple locations if
the first location recieved is invalid.
* Cleaned up inc_visiting removing redundant functions that were used by
the webapp.

Requires a new group object RVSDF6A3CYU2GY4 in eleven-fixtures to work.

Fixes https://trello.com/c/0Gwh0JFc