Make location equality check work with objref proxies
Checks whether the player is in their home street/house should work
regardless of implementation details (i.e. whether the POL reference
points to the actual game object or an objref). With the GS change for
random POL/GS distribution, this is required to make the butler box
behave correctly in all cases.
The change in `croak` is just an equivalent adjustment; all other
similar equality tests are already based on comparing TSIDs.